memtest: Modify memtest
Modify memtest in which the greatest change is moving from getting the
virtual address of the extender area from platform_data populated from
the intel_extender (parent) into ioremap() intercepting the calls and
returning the extender address to the caller. In other words the
extender client is totally unaware it uses the extender area with the
control only in device-tree.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>