extender mapper: Add in important updates that are yet in progress
Go for the break-before-make for page table allocation
https://patchwork.kernel.org/project/linux-arm-kernel/patch/1461856591-5751-1-git-send-email-marc.zyngier@arm.com/

From another angle, this commit tries to resolve a twofold situation:
multiple faults arrive, only one allowed in, the others wait for acquiring
the lock and after let in make the same changes to page table entries.
Also it has been observed that the page table entries getting allocated are
not only populated already but hold a different value, meaning that
probably one VA mapped to two PAs that is of course illegal.
(Note, two VAs mapping to the same PA is valid)

Summing up the two resolutions are proposed: break-before-make and
upon PG entires same return.

Noting secondly this is work-in-progress and calls for more work!

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>