extender el0: Unmap with use of vma in the window struct.
It makes sense to use it instead of the task but for a reason it
oopses in "pc : find_vma+0x4c/0xa0". Probably at the time the vma is
processed the task with mm gets torn down.

I'm storing the commit to be able to get back and work more.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>