extender el0 and el1: Add extender trace call
Extender trace call may seem not intuitive at first sight but should be of
use. We should define the number of frames to print the function names.
For the first function we can pass, but we can omit as well, the argument
variables the function is called with. So that for it to make sense you
should inspect the function arguments, pick ones of interest and provide
it to the tracer - I found it useful.

As an example the tracer below will print 5x function up from current
one. Also as the arguments, VA_ARGS, are provided it will print them
along with the function name.

In code:
extender_trace_call(5, "vma->vm_start %lx vma->vm_end %lx pfn %lx vma->vm_pgoff %lx\n",
                    vma->vm_start, vma->vm_end, pfn, vma->vm_pgoff);

On system printing:
extender trace call:
1: rdma_user_mmap_io+0x0/0x1a0 [ib_uverbs] args: vma->vm_start ffff985b5000 \
	vma->vm_end ffff985b6000 pfn 2012089 vma->vm_pgoff 7
2: mlx5_ib_mmap+0x1bc/0x3b0 [mlx5_ib]
3: ib_uverbs_mmap+0x84/0xd4 [ib_uverbs]
4: mmap_region+0x40c/0x654
5: do_mmap+0x368/0x46c

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>