extender el0: RDMA works!
First RDMA tests seem to work. There are loads of things to code differently,
that will be updates in the commits to come. It is yet more exercise than
'feature' implemented.

Also use the upper nibble 'signalization' for RDMA to command it to use
the extender (aka dynamic windows) instead of static ones.

To recall, upper nibble 'signalization' makes use of the 47th bit set to
the physical address written to dt.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>