extender and RDMA: Make use (yet) upper nibble hack
Until this commit if RDMA touse (or not) the ASE windows dynamically was
manged in the code. With this commit we make it decide in the device-tree
by setting the upper nibble (exactly as we have it to now for el1).

The upper nibble hack is to be removed as it asks for trouble sooner or
later. It is not that it is ugly (thought it is) but it is also that
if there is any Master but CPU using the physical address hacked
(47th bit set) it will go then utterly wrong.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>