extender el0 and el1: Expose some of the window properties via device attributes
Enable some of the properties to user-space via sysfs. The windows have a
number of them, but because of the limited (and not readable) output if
we tried to print all of them we describe the windows by two of them, their
number and physical start address, namely num:phys_addr.

To recall, windows can be on one of the four lists, allocated or free running
either el0 or el1 giving four lists all together, therefore there should
be four types of attributes available.

Running the following on the command line:

  d=`find /sys/devices/platform -name "*extender"`; cd $d; ar=(el*)
  for list in ${ar[@]}; do echo "$list > $(cat $list)"; done

should output similar to:

  el0_allocated > 31:201f000000 30:201e000000
  el0_free >
  el1_allocated > 7:2007000000 6:2006000000 5:2005000000 4:2004000000
  el1_free > 29:201d000000 28:201c000000 27:201b000000 26:201a000000

It can be read that el0 has two windows numbered 31, and 30, both
allocated, and el1 has eight windows, 7, 6, 5, 4 allocated and 29, 28,
27 and 26 free (aka available for allocation when the CPU faults).

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>