extender: Fix for using the AT (Address translation) for el1
May be that a faulting addr is already resolved. A thread with
the same address could be busy waiting acquiring the lock
while within the critical section (mapping path) the addr could
have been already resolved. In that case don't re-resolve aka
re-map the address. Note though it will be no harm doing so but
would unnecessary undermine the performance.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>