memtest: Relax memtest
memtest originally was walking through the available rams with a 4-byte
step writing and reading back 1000 times. That was making the scheduler
unhappy leading to CPU stalls.

Relax the memtest in which it walks the rams with a 1K step resulting
with four writes and reads going over the rams only once.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>